Implement function AI_WhirlAround
It is supposed to be a faster version of AI_TurnToNpc and seemlingly
used in Gothic 1 in a number of places.

The script does not contain specific values for whirl-speed, so just use
double of the NPc's turn-speed.

This looked quite nice with the guards at the New Camp when drawing a
weapon behind them.