Implement Npc_GetInvItemBySlot script function
This seems mainly used by Gothic 1 with its categorized inventory and is
for example used by Baal Lukor in the Ork graveyard to find a weapon in
his inventory to defend against the Orks.

The implementation just iterates over the sorted inventory and finds
the Nth item of the given category, essentially emulating the item-slot
thing.