Handle Gothic1 guilds
The numeric values of the Gothic 1 guilds are different from Gothic 2.
While for the most part this does not matter (the divider between
humans and monsters is unchanged).

The one field where it matters now, is the check if a spell is applicable
to an enemy, like killing only the undead.

So add the G1 constants and add a separate check if running a Gothic 1.