Implement function ai_turnaway
Defined as 'npc' turns away from 'other', just like turnTo
but +180 degrees.

One user is for example Baal Netbek in the swamp where the player
character ends the dialog with turning away saying "This guy won't
be of help", the other example is the fire demon in Xardas' tower,
where the player first turns away before whirling back to the demon.