Protect the renderer against not finding any valid shadow or zbuffer formats
While unlikely, there may be cases with broken or not yet feature-complete
graphics drivers, where the renderer might not find valid formats for
shadows or zbuffers.

When that happens the default values will stay in place and the game will
fail later with very obscure exceptions.

The TextureFormat enum in Tempest already has a "Last" element, so use
that as a sentinel to catch the no-valid-format case.

As there is no way to recover from this and it shouldn't happen in the
first place, throw an exception.