Prevent deadlocking the perceptionProcess when perceptiontime is not set
It seems some Npc do not get a perception-time set, which lets it stay
at 0. If the perceptionProcess runs with that value, it will immediately
rerun itself again, potentially leading to a deadlock.

The Npc where this happens in practice is the Firedemon in Xardas' tower
in Gothic 1. When the player reaches it, all Npc motion will stop and
with enabled logging, everthing will be flooded with messages about the
Firedemon perception process - rerunning the AssessPlayer routine over
and over.