Add whirl animation types
All characters (except only the humans in Gothic 2) have T_SURPRISE_CW
and T_SURPRISE_CCW animations used in the whirl-around type of movement.

This is mainly used in Gothic 1 to portray a different style of turning
to an enemy or surprising event.

This adds the animation to the solver. Normally the whirl animation should
not happen in water, but keep the case around just to be on the save side.