Try a MDL resource as alternative for script resources (#716)
When a MDS file is loaded, the resource manager also needs to load the mesh
and hirarchy. So far these always come from MDH and MDM files that contain
both elements separately.

In some data files, MDL files are used instead that contain both components
in the same file. So add a fallback to create the ProtoMesh from that.