Don't set victim when picking up an item from the floor (#735)
At least the Gothic 1 script assesses theft differently depending on whether
a victim is set or not.

if (npc_canseenpc(self, other)) {
        printdebugnpc(pd_zs_check, "...NSC kann den Dieb sehen!");
        if (hlp_isvaliditem(item) && (!hlp_isvalidnpc(victim))) {
            printdebugnpc(pd_zs_check, "...Item wurde aufgehoben!"); // Item picked up
            ...
        }
        else {
            printdebugnpc(pd_zs_check, "...Taschendiebstahl!"); //pickpocket
            ...
        }

This causes NPC around an item to tell the player that something is getting
stolen right now and also indicates that victim is not always set in the
perception process.

So adapt the sent passivePerc accordingly to not transmit the player as victim
when picking up items.

GameScript::invokeState() already removed the default setting of other=npc
for the !other case in commit 3fc5c39e2095 ("damage from faling") as part
of the fix for #18, so do the same for the current default victim setting.

This fixes #717.