2 additional AI turn actions (#718)
* Add whirl animation types

All characters (except only the humans in Gothic 2) have T_SURPRISE_CW
and T_SURPRISE_CCW animations used in the whirl-around type of movement.

This is mainly used in Gothic 1 to portray a different style of turning
to an enemy or surprising event.

This adds the animation to the solver. Normally the whirl animation should
not happen in water, but keep the case around just to be on the save side.

* Move turn-preparation to a separate function

There are multiple turning variants (regular, whirl, away) that all will
need the same preparation steps, so move them to a separate function
for re-use.

* Implement function AI_WhirlAround

It is supposed to be a faster version of AI_TurnToNpc and seemlingly
used in Gothic 1 in a number of places.

It's for example used when the player draws a weapon behind an NPC
but also with the Firedaemon in Xardas' tower.

* Implement function ai_turnaway

Defined as 'npc' turns away from 'other', just like turnTo
but +180 degrees.

One user is for example Baal Netbek in the swamp where the player
character ends the dialog with turning away saying "This guy won't
be of help", the other example is the fire demon in Xardas' tower,
where the player first turns away before whirling back to the demon.