fix unintenional weapon unequip
Ranged weapon auto equip would accidentally unequip current equipped weapon if
no best weapon could be found.

In Cavalorn case this is because script function EquipItem doesn't check ammo
availability but the AI_BestWeapon functions do. So the weapon is first
equipped and later unequipped because no best weapon could be found.

The fix is to just do nothing in this case. For consistency this has been
added for armor and melee weapons too but should never be triggered.