fix unwanted npc removal
Regression introduced in b7f8ed.

Problem was npcs with invalid waypoints are removed in
Npc::resetPositionToTA.

Vanilla checks the current and next routine and would spawn npc only if
one of these has a valid waypoint. But npc would still appear sometime
later if there's a valid routine and is then never removed from the
world. This needs more investigation and is subject for a future PR.