fix fleeing for G1 npcs (#824)
G1 script makes npcs flee if the their health drops below a certain
percentage in a fight. As in many ZS loop functions there's a AI_Wait
call in ZS_Flee.

The Engine is calling Npc::adjustAttackRotation function everytime a npc
has to wait. Because fleeing npcs have a target and a drawn weapon a
rotation is executed. The rotation added from AI_Flee is overwritten and
the npc follows the player instead of running away.

Don't execute Npc::adjustAttackRotation in that case.